+++ auth-backend-api/src/main/java/com/juanfedevmaster/authbackendapi/services/IProfileService.java
@@ -6,7 +6,7 @@ import com.juanfedevmaster.authbackendapi.entity.dto.ProfileResponse;
 public interface IProfileService {
     ProfileResponse createProfile(ProfileRequest request);
     ProfileResponse getProfileById(Integer id);
-    ProfileResponse getProfileByUserId(Integer userId);
+    ProfileResponse getProfileByUserId(Long userId);
     ProfileResponse updateProfile(Integer id, ProfileRequest request);
     void deleteProfile(Integer id);
 }
